ts migration done, code needs refactor
