+++ App/package.json
@@ -3,12 +3,17 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@types/jest": "^29.5.2",
+    "@types/node": "^20.3.3",
+    "@types/react": "^18.2.14",
+    "@types/react-dom": "^18.2.6",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-map-gl": "5.3.19",
     "react-router-dom": "^6.2.1",
     "react-scripts": "5.0.0",
-    "sweetalert2": "^11.3.4"
+    "sweetalert2": "^11.3.4",
+    "typescript": "^5.1.6"
   },
   "scripts": {
     "start": "react-scripts start",
