+++ App/src/Components/Home/Homepage.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+import Map from "../Map/Map";
+
+export const Homepage = () => (
+    <div>
+        <Map />
+    </div>
+);
+
+export default Homepage;
