--- App/src/Components/Homepage.jsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import Map from './Map';
-
-
-export default function Homepage({ controllerVal }) {
-
-	return (
-		<div>
-			<Map controllerVal={controllerVal} />
-		</div>
-	);
-}
