+++ App/src/Components/Login/Login.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function Login() {
+    return <div style={{ zIndex: "1" }} className="col-3 ms-3"></div>;
+}
+
+export default Login;
