+++ App/src/Constants/constants.ts
@@ -1,0 +1,11 @@
+export const CONSTANTS = {
+    mapStyleURL: `mapbox://styles/jayantmanchanda/ckynbop6g56nw14l2m5r2t7bd`,
+    info: {
+        icon: "info",
+        text: "I am currently working on setting up a process for this",
+    },
+    process: {
+        icon: "What should i do ?",
+        text: "By interacting with a map, you can explore various locations marked by pink dots. Simply click on a dot, and a popup will appear, providing details about that specific place. If you wish to immerse yourself further, click on the microphone icon. Your browser will prompt you to grant microphone permissions, enabling you to hear your own voice within the selected location. For optimal experience, remember to use headphones to prevent any audio feedback loops.",
+    },
+};
