+++ App/src/helpers/swal.ts
@@ -1,0 +1,5 @@
+import * as Swal from "sweetalert2";
+
+export const invokeSwal = (toInvoke: {}): void => {
+    (Swal as any).fire(toInvoke);
+};
